+++ operators/advanced_network/advanced_network/manager.h
@@ -45,9 +45,9 @@ class Manager {
 
   // Common free functions to override
   virtual void* get_packet_ptr(BurstParams* burst, int idx) = 0;
-  virtual uint16_t get_packet_length(BurstParams* burst, int idx) = 0;
+  virtual uint32_t get_packet_length(BurstParams* burst, int idx) = 0;
   virtual void* get_segment_packet_ptr(BurstParams* burst, int seg, int idx) = 0;
-  virtual uint16_t get_segment_packet_length(BurstParams* burst, int seg, int idx) = 0;
+  virtual uint32_t get_segment_packet_length(BurstParams* burst, int seg, int idx) = 0;
   virtual uint16_t get_packet_flow_id(BurstParams* burst, int idx) = 0;
   virtual void* get_packet_extra_info(BurstParams* burst, int idx) = 0;
   virtual Status get_tx_packet_burst(BurstParams* burst) = 0;
